Added purple text class for Eric's name in index.html and imported it in index.css
This commit adds a new CSS class called ".purple-text" to style the name "Eric" in purple in the index.html file. The class is defined in the index.css file, which has also been updated to include the new class. The name "Eric" now appears in purple on the webpage.